requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2

annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
aurelio-sdk==0.0.18
backports.tarfile==1.2.0
black==24.8.0
blis==0.7.11
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
cohere==5.14.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.4
contourpy==1.3.0
crashtest==0.4.1
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.12
fastavro==1.10.0
fastjsonschema==2.19.1
filelock==3.13.1
findpython==0.6.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.2.0
elevenlabs
together==1.4.6
langchain_cohere
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
groq==0.19.0
gspread==6.1.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.5.2
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.4
ipython==8.18.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
jovian==0.2.47
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.7
langchain-community==0.3.19
langchain-core==0.3.45
langchain-tests==0.3.14
langchain-text-splitters==0.3.6
langcodes==3.4.0
langsmith==0.3.15
language_data==1.2.0
langchain-huggingface
langchain-groq
langchain-pinecone
langgraph
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.2.0
openai==1.66.3
opencv-python==4.11.0.86
openpyxl==3.1.2
orjson==3.10.15
packaging==24.0

parso==0.8.4
pathspec==0.12.1
pbs-installer==2025.3.17
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pillow
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pkginfo==1.12.1.2
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
preshed==3.0.9
prompt-toolkit==3.0.43
propcache==0.3.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.0.0
shellingham==1.5.4
langgraph-checkpoint-postgres
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.4.8
stack-data==0.6.3
starlette==0.46.1
sympy==1.12
syrupy==4.9.0
tenacity==8.3.0
thefuzz==0.22.1
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.20.1
tomli==2.0.1
tomlkit==0.13.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.2
trove-classifiers==2025.3.19.19
typer==0.9.4
types-PyYAML==6.0.12.20241230
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
ultralytics==8.3.40
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==1.26.20
uuid==1.30
uvicorn==0.34.0
virtualenv==20.30.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websockets==15.0.1
xattr==1.1.4
yarl==1.18.3
zipp==3.18.2
zstandard==0.23.0
psycopg[binary]